Fix deprecated casts in loop indices
All enumerating loops index by size_t or const size_t.
The body of the loop is responsible for explicit conversion to int.
Methods should still be written to accept int, not size_t.

+++ src/physics/tribe.d
@@ -166,7 +166,7 @@ final class Tribe {
             import std.string;
             string msg = format("Style %s spawns from wrong hatch #%d.",
                 this.style, this.nextHatch);
-            foreach (int i, hatch; ow.state.hatches) {
+            foreach (long i, hatch; ow.state.hatches) {
                 msg ~= format("\nHatch #%d has styles:", i);
                 foreach (Style st; hatch.tribes) {
                     msg ~= " " ~ styleToString(st);
